+++ pyswtools/main.py
@@ -4,28 +4,29 @@ Docstring
 
 import click
 
+from .config import config
 from .ready_dxf.main import ready_dxf
+from .copy_full_assembly.main import copy_full_assembly
 
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 __author__ = "Devillez Louis, Kjell Magne Fauske"
 __maintainer__ = "Deville Louis"
 __email__ = "louis.devillez@gmail.com"
 
 
-@click.group(invoke_without_command=True)
+@click.group()
 @click.version_option(__version__, "-v", "--version")
 @click.help_option("-h", "--help")
 def cli() -> None:
     """
     Combination of commands to help you work with solidworks
     """
-    ctx = click.get_current_context()
-    click.echo(ctx.get_help())
-    ctx.exit()
 
 
+cli.add_command(config)
 cli.add_command(ready_dxf)
+cli.add_command(copy_full_assembly)
 
 if __name__ == "__main__":
     cli()